Move RpcDistributionStats from xds_url_map_testcase to grpc_testing
